FinishedServicesRelationship

+++ Services/CustomerService/Models/ItemList.cs
@@ -2,13 +2,11 @@ namespace CustomerService.Models
 {
     public class ItemList
     {
-        public int ItemListId { get; set; } // Primary Key
-        public int CartId { get; set; } // Foreign Key
-        public int ProductId { get; set; } // From Inventory Service
+        public int ItemListId { get; set; } 
+        public int CartId { get; set; } 
+        public int ProductId { get; set; } 
         public int Quantity { get; set; }
-
-        public float UnitPrice { get; set; } 
-
+        public decimal UnitPrice { get; set; } 
         public CartInProgress? Cart { get; set; }
     }
 }
